% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllLeaves.R
\name{AllLeaves}
\alias{AllLeaves}
\title{AllLeaves gives all choices of leaves in tree}
\usage{
AllLeaves(L, k)
}
\value{
Output shows all choices of leaves in tree
}
\description{
AllLeaves gives all choices of leaves in tree
}
\details{
L is maximal level of the tree, k is number of options (number of alphabet)
}

