% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fibonacci.R
\name{Fibonacci}
\alias{Fibonacci}
\title{Function Fibonacci is funtion which takes the value of n-th element from Fibonacci numbers}
\usage{
Fibonacci(n)
}
\value{
n-th element from Fibonacci numbers
}
\description{
Function Fibonacci is funtion which takes the value of n-th element from Fibonacci numbers
}

