% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hanojskavez.R
\name{Hanojskavez}
\alias{Hanojskavez}
\title{Hanojskavez is function which displays solution Tower of Hanoi.}
\usage{
Hanojskavez(n, zkoliku, nakolik)
}
\description{
Hanojskavez is function which displays solution Tower of Hanoi.
}
\details{
n is number of disks; zkoliku is place from are disks moving, nakolik is place, where should be every disks in right order.
}

