% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hirschberg.R
\name{Hirschberg}
\alias{Hirschberg}
\title{Hirschberg function return align sequences. This align is made by Hirschberg algorithm.}
\usage{
Hirschberg(X, Y, Align, match, mismatch, gap)
}
\description{
Hirschberg function return align sequences. This align is made by Hirschberg algorithm.
}

