% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HM.R
\name{HM}
\alias{HM}
\title{HM is function which return probability, that sequence is CpG island. If return>0 it isn't CpG island, if return>0 it is CpG island.}
\usage{
HM(sek)
}
\description{
HM is function which return probability, that sequence is CpG island. If return>0 it isn't CpG island, if return>0 it is CpG island.
}
\details{
input is part of sequence, where we wanted to know, if it should be the CpG island
}

