% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCP_vypocet.R
\name{LCP_vypocet}
\alias{LCP_vypocet}
\title{Function LCP_vypocet return vector, where value lcp[i] indicates length of the longest common prefix of the suffixes inexed by suffix[i] and suffix[i+1]}
\usage{
LCP_vypocet(slovo)
}
\description{
Function LCP_vypocet return vector, where value lcp[i] indicates length of the longest common prefix of the suffixes inexed by suffix[i] and suffix[i+1]
}

