% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NextLeaf.R
\name{NextLeaf}
\alias{NextLeaf}
\title{NextLeaf demonstrates how to jump from an L-mer a = (a1 a2 ··· aL) to the next L-mer in the progression.}
\usage{
NextLeaf(a, L, k)
}
\value{
new position where we are
}
\description{
NextLeaf demonstrates how to jump from an L-mer a = (a1 a2 ··· aL) to the next L-mer in the progression.
}
\details{
a is position where we are; L is maximal level of the tree, k is number of options (number of alphabet)
}

