% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NextVertex.R
\name{NextVertex}
\alias{NextVertex}
\title{NextVertex function moves down to the next lower level and explores that subtree of a.}
\usage{
NextVertex(a, i, L, k)
}
\value{
new position where we are and level where we are
}
\description{
NextVertex function moves down to the next lower level and explores that subtree of a.
}
\details{
a is position where we are;i is level where we are ;L is maximal level of the tree, k is number of options (number of alphabet)
}

