% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NWScore.R
\name{NWScore}
\alias{NWScore}
\title{Function NWScore return the last row of scoring matrix creating by  Needleman-Wunsch scoring system.}
\usage{
NWScore(Sek1, Sek2, match, mismatch, gap)
}
\description{
Function NWScore return the last row of scoring matrix creating by  Needleman-Wunsch scoring system.
}
\details{
sek1 and sek2 are the sekvence of DNA, which we wonted align
}

