% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartialDigestProblem.R
\name{PartialDigestProblem}
\alias{PartialDigestProblem}
\title{PartialDigestProblem function given the (multi)set ∆X, compute a set X which could have produced ∆X in restriction mapping.}
\usage{
PartialDigestProblem(L)
}
\value{
vector of restriction place
}
\description{
L is vector of ∆X (fragments vector)
}

