% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suffix_array.R
\name{suffix_array}
\alias{suffix_array}
\title{suffix_array is function giving every suffix, suffix array as string and in vector of order, finally invers suffyx array in vector order.}
\usage{
suffix_array(slovo)
}
\description{
suffix_array is function giving every suffix, suffix array as string and in vector of order, finally invers suffyx array in vector order.
}
\details{
input is sequence; outputs: suffix conteins all posible suffixes, abeceda is an array of all suffixes of A in lexicographical order, poradi is vector of order abeceda, inverz is invers suffix array
}

