% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VratitMince.R
\name{VratitMince}
\alias{VratitMince}
\title{Function VratitMince takes vector which number of coins should be get back.}
\usage{
VratitMince(M, c)
}
\value{
number of coins which should be get back
}
\description{
Function VratitMince takes vector which number of coins should be get back.
}
\details{
M is amount of money which should be get back; c is vector of value coins, which should be used
}

